refactor: move functions into util files
