+++ src/index.ts
@@ -1,1 +1,1 @@
-export * from './decorators/ApiException';
+export * from './lib';
