+++ src/lib/decorators/api-exception.decorator.spec.ts
@@ -4,7 +4,7 @@ import { BadRequestException, NotFoundException } from '@nestjs/common';
 import { ApiOperation, ApiOperationOptions, ApiResponse, ApiResponseOptions } from '@nestjs/swagger';
 import { DECORATORS } from '@nestjs/swagger/dist/constants';
 
-import { buildTemplatedApiExceptionDecorator } from './ApiException';
+import { buildTemplatedApiExceptionDecorator } from '../';
 
 const TemplatedApiException = buildTemplatedApiExceptionDecorator({
   statusCode: '$status',
