+++ src/lib/interfaces/exceptions.interface.ts
@@ -1,4 +1,4 @@
 import { HttpException, Type } from '@nestjs/common';
 
 export type Exception<T extends HttpException> = Type<T> | T;
-export type ExceptionsArguments<T extends HttpException> = Exception<T> | Exception<T>[];
+export type ExceptionArguments<T extends HttpException> = Exception<T> | Exception<T>[];
